renamed saveOrUpdate() to addOrUpdate() + renamed remote() to http()
